+++ argopy/stores/argo_index_proto.py
@@ -3,6 +3,7 @@ Argo file index store prototype
 
 """
 
+import os
 import numpy as np
 import pandas as pd
 import logging
@@ -493,6 +494,10 @@ class ArgoIndexStoreProto(ABC):
 
         return df
 
+    def to_indexfile(self):
+        """Save search results on file, following the Argo standard index formats"""
+        raise NotImplementedError("Not implemented")
+
     @property
     @abstractmethod
     def search_path(self):
@@ -697,3 +702,24 @@ class ArgoIndexStoreProto(ABC):
 
         """
         raise NotImplementedError("Not implemented")
+
+    def _insert_header(self, originalfile):
+        header = """# Title : Profile directory file of the Argo Global Data Assembly Center
+# Description : The directory file describes all individual profile files of the argo GDAC ftp site.
+# Project : ARGO
+# Format version : 2.0
+# Date of update : %s
+# FTP root number 1 : ftp://ftp.ifremer.fr/ifremer/argo/dac
+# FTP root number 2 : ftp://usgodae.org/pub/outgoing/argo/dac
+# GDAC node : CORIOLIS
+file,date,latitude,longitude,ocean,profiler_type,institution,date_update
+""" % pd.to_datetime('now', utc=True).strftime('%Y%m%d%H%M%S')
+
+        with open(originalfile, 'r') as f:
+            data = f.read()
+
+        with open(originalfile, 'w') as f:
+            f.write(header)
+            f.write(data)
+
+        return originalfile